requirements.txt
absl-py==2.0.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.1
Flask==3.0.0
fsspec==2023.12.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.2
Pillow==9.5.0
protobuf==4.23.4
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyre-extensions==0.0.30
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
torch==2.0.0
torcheval==0.0.6
torchtnt==0.0.7
torchvision==0.15.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
